requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
tensorflow-gpu==2.10.0
flask==2.3.2
easyocr==1.7.0
pycocotools==2.0.6
albumentations==1.3.0
scikit-learn==1.2.2
pillow==9.5.0
tkmacosx==1.0.5
gradio==3.34.0